third_party/blink/web_tests/storage/indexeddb/index-get-key-argument-required-expected.txt

Test IndexedDB index .get() required argument

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "index-get-key-argument-required.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
objectStore = db.createObjectStore('foo', { keyPath: 'id', autoIncrement: true });
index = objectStore.createIndex('first', 'first');
PASS index.get(); threw exception TypeError: Failed to execute 'get' on 'IDBIndex': 1 argument required, but only 0 present..
PASS index.getKey(); threw exception TypeError: Failed to execute 'getKey' on 'IDBIndex': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
